[REFACTOR] #9 - NetWorthTooltipActivity 패키지 이동
