+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.SpoonFeed">
         <activity
-            android:name=".NetWorthTooltipActivity"
+            android:name=".presentation.ui.policy.view.filter.tooltip.NetWorthTooltipActivity"
             android:exported="true" />
         <activity
             android:name=".presentation.ui.policy.view.filter.tooltip.MedianIncomeTooltipActivity"
